+++ src/main/java/project/linkarchive/backend/link/controller/LinkQueryController.java
@@ -12,9 +12,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import project.linkarchive.backend.bookmark.service.BookMarkQueryService;
 import project.linkarchive.backend.link.response.LinkMetaDataResponse;
-import project.linkarchive.backend.link.response.RefactorUserLinkList.UserLinkListResponse;
+import project.linkarchive.backend.link.response.UserLinkList.UserLinkListResponse;
 import project.linkarchive.backend.link.response.linkList.UserExcludedLinkListResponse;
 import project.linkarchive.backend.link.service.LinkQueryService;
 import project.linkarchive.backend.security.AuthInfo;
