+++ src/main/java/project/linkarchive/backend/link/response/UserLinkList/LinkResponse.java
@@ -1,11 +1,10 @@
-package project.linkarchive.backend.link.response.RefactorUserLinkList;
+package project.linkarchive.backend.link.response.UserLinkList;
 
+import com.querydsl.core.annotations.QueryProjection;
 import lombok.Getter;
 
-import java.util.List;
-
 @Getter
-public class UserLinkResponse {
+public class LinkResponse {
 
     private Long urlId;
     private String link;
@@ -13,16 +12,16 @@ public class UserLinkResponse {
     private String description;
     private String thumbnail;
     private Long bookMarkCount;
-    private List<TagResponse> tagList;
 
-    public UserLinkResponse(Long urlId, String link, String title, String description, String thumbnail, Long bookMarkCount, List<TagResponse> linkTagList) {
+    @QueryProjection
+    public LinkResponse(Long urlId, String link, String title, String description, String thumbnail, Long bookMarkCount) {
         this.urlId = urlId;
         this.link = link;
         this.title = title;
         this.description = description;
         this.thumbnail = thumbnail;
         this.bookMarkCount = bookMarkCount;
-        this.tagList = linkTagList;
+
     }
 
 }
