+++ src/main/java/project/linkarchive/backend/link/response/UserLinkList/UserLinkListResponse.java
@@ -1,4 +1,4 @@
-package project.linkarchive.backend.link.response.RefactorUserLinkList;
+package project.linkarchive.backend.link.response.UserLinkList;
 
 import lombok.Getter;
