+++ src/main/java/project/linkarchive/backend/link/service/LinkQueryService.java
@@ -10,10 +10,10 @@ import project.linkarchive.backend.hashtag.response.TagListDetailResponse;
 import project.linkarchive.backend.link.domain.UrlHashTag;
 import project.linkarchive.backend.link.repository.UrlHashTagRepository;
 import project.linkarchive.backend.link.repository.UrlRepositoryImpl;
-import project.linkarchive.backend.link.response.RefactorUserLinkList.LinkResponse;
-import project.linkarchive.backend.link.response.RefactorUserLinkList.TagResponse;
-import project.linkarchive.backend.link.response.RefactorUserLinkList.UserLinkListResponse;
-import project.linkarchive.backend.link.response.RefactorUserLinkList.UserLinkResponse;
+import project.linkarchive.backend.link.response.UserLinkList.LinkResponse;
+import project.linkarchive.backend.link.response.UserLinkList.TagResponse;
+import project.linkarchive.backend.link.response.UserLinkList.UserLinkListResponse;
+import project.linkarchive.backend.link.response.UserLinkList.UserLinkResponse;
 import project.linkarchive.backend.link.response.linkList.UserExcludedLinkListDetailResponse;
 import project.linkarchive.backend.link.response.linkList.UserExcludedLinkListResponse;
 import project.linkarchive.backend.link.response.linkList.UserExcludedLinkTagListDetailResponse;
